requirements.txt
scapy
slack-sdk
python-nmap
yaml
